Hooked up pocketnes to the front end, with some final refactoring
